Title: A/B 출력, Time: 132ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 251111/A/B 출력/a-divide-b.java
@@ -6,6 +6,6 @@ public class Main {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        System.out.printf("%.21f",a/b);
+        System.out.printf("%.21f",(double)a/b);
     }
 }
